+++ tests/parser/token.cc
@@ -9,8 +9,8 @@ SCENARIO("Create some tokens and check their values") {
     Token token(TokenType::VALUE, "identifier1.identifier2", 24, 456);
     REQUIRE(token.getType() == TokenType::VALUE);
     REQUIRE(token.getValue() == "identifier1.identifier2");
-    REQUIRE(token.getStartLineNo() == 24);
-    REQUIRE(token.getStartCharNo() == 456);
+    REQUIRE(token.getStartLineNo() == (unsigned long long int)24);
+    REQUIRE(token.getStartCharNo() == (unsigned long long int)456);
   }
 
   /// Check that a value token works as expected with just an identifier
@@ -18,8 +18,8 @@ SCENARIO("Create some tokens and check their values") {
     Token token(TokenType::VALUE, "component_Name", 5427354, 65474);
     REQUIRE(token.getType() == TokenType::VALUE);
     REQUIRE(token.getValue() == "component_Name");
-    REQUIRE(token.getStartLineNo() == 5427354);
-    REQUIRE(token.getStartCharNo() == 65474);
+    REQUIRE(token.getStartLineNo() == (unsigned long long int)5427354);
+    REQUIRE(token.getStartCharNo() == (unsigned long long int)65474);
   }
 
   /// Check that a string token works as expected
@@ -27,8 +27,8 @@ SCENARIO("Create some tokens and check their values") {
     Token token(TokenType::STRING, "string, this is", 1, 1);
     REQUIRE(token.getType() == TokenType::STRING);
     REQUIRE(token.getValue() == "string, this is");
-    REQUIRE(token.getStartLineNo() == 1);
-    REQUIRE(token.getStartCharNo() == 1);
+    REQUIRE(token.getStartLineNo() == (unsigned long long int)1);
+    REQUIRE(token.getStartCharNo() == (unsigned long long int)1);
   }
 
   /// Check that a dict opening delimiter token works as expected
@@ -36,8 +36,8 @@ SCENARIO("Create some tokens and check their values") {
     Token token(TokenType::DICTDELIMOPEN, "{", 99, 546);
     REQUIRE(token.getType() == TokenType::DICTDELIMOPEN);
     REQUIRE(token.getValue() == "{");
-    REQUIRE(token.getStartLineNo() == 99);
-    REQUIRE(token.getStartCharNo() == 546);
+    REQUIRE(token.getStartLineNo() == (unsigned long long int)99);
+    REQUIRE(token.getStartCharNo() == (unsigned long long int)546);
   }
 
   /// Check that a dict closing delimiter token works as expected
@@ -45,35 +45,35 @@ SCENARIO("Create some tokens and check their values") {
     Token token(TokenType::DICTDELIMCLOSE, "}", 777, 888);
     REQUIRE(token.getType() == TokenType::DICTDELIMCLOSE);
     REQUIRE(token.getValue() == "}");
-    REQUIRE(token.getStartLineNo() == 777);
-    REQUIRE(token.getStartCharNo() == 888);
+    REQUIRE(token.getStartLineNo() == (unsigned long long int)777);
+    REQUIRE(token.getStartCharNo() == (unsigned long long int)888);
   }
 
   /// Check that a dict separator token works as expected
   /// Check maximum value boundary case for positions
   WHEN("Create a dictionary separator token") {
     Token token(TokenType::DICTSEPARATOR, ",",
-                std::numeric_limits<unsigned int>::max(),
-                std::numeric_limits<unsigned int>::max());
+                std::numeric_limits<unsigned long long int>::max(),
+                std::numeric_limits<unsigned long long int>::max());
     REQUIRE(token.getType() == TokenType::DICTSEPARATOR);
     REQUIRE(token.getValue() == ",");
     REQUIRE(token.getStartLineNo() ==
-      std::numeric_limits<unsigned int>::max());
+      std::numeric_limits<unsigned long long int>::max());
     REQUIRE(token.getStartCharNo() ==
-      std::numeric_limits<unsigned int>::max());
+      std::numeric_limits<unsigned long long int>::max());
   }
 
   /// Check that a pair separator token works as expected
   /// Check minimum value boundary case for positions
   WHEN("Create a pair separator token") {
     Token token(TokenType::PAIRSEPARATOR, ":",
-                std::numeric_limits<unsigned int>::min(),
-                std::numeric_limits<unsigned int>::min());
+                std::numeric_limits<unsigned long long int>::min(),
+                std::numeric_limits<unsigned long long int>::min());
     REQUIRE(token.getType() == TokenType::PAIRSEPARATOR);
     REQUIRE(token.getValue() == ":");
     REQUIRE(token.getStartLineNo() ==
-      std::numeric_limits<unsigned int>::min());
+      std::numeric_limits<unsigned long long int>::min());
     REQUIRE(token.getStartCharNo() ==
-      std::numeric_limits<unsigned int>::min());
+      std::numeric_limits<unsigned long long int>::min());
   }
 }
